requirements.txt
Flask 
pillow
freeGPT

